+++ test/bag/make-bag_test.go
@@ -28,3 +28,13 @@ func TestMakeBagSet(t *testing.T) {
 		Expect: `"{a: 7}"`,
 	}).Test(t)
 }
+
+func TestMakeBagBag(t *testing.T) {
+	(&sliptest.Function{
+		Source: `(let* ((b1 (make-bag "{a:1}"))
+                        (b2 (make-bag b1)))
+                   (bag-set b1 2 "b")
+                   (list (bag-write b1 nil) (bag-write b2 nil)))`,
+		Expect: `("{a: 1 b: 2}" "{a: 1}")`,
+	}).Test(t)
+}